requirements.txt
dj-database-url
django_heroku
gunicorn
psycopg2
pytz
whitenoise
django-contrib-comments
Pillow
django-session-timeout
django-storages
django-background-tasks
pygooglevoice
django-phone-field
requests
django-user-visit==0.4
django-allauth

